fst_backend/requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
pillow==11.1.0
pip==24.0
PyJWT==2.9.0
pymongo==4.11.3
sqlparse==0.5.3
